# A couple small perf improvements