A couple small changes to improve tutorial loading perf. All of these measurements were done in chrome with the 6x CPU slowdown:

1. When rendering the markdown snippets, do it outside the dom to prevent a style recalculation (went from ~90ms to ~45ms)
2. Replaced the UTF8 encoding/decoding we do in a few places to use the TextEncoder/TextDecoder APIs which are now supported in all the browsers we support and node. I mainly did this because I was noticing a lot of time spent getting a SHA256 hash on extensions (went from ~150ms to ~60ms but this is called a few times with different timings)

AFAIK the result of the TextEncoder/TextDecoder should be the same as our implementations. I confirmed it at least returned the same results in the calls to the SHA256 function.

Doesn't make a *ton* of difference on my machine but I can see the improvements in the performance tab of devtools. Hopefully I will have time to go into the office and pick up a chromebook this week (my work laptop is a little too speedy for accurate numbers).