@pelikhan 2021-09-21
saved by node 16.0 https://nodejs.org/api/util.html#util_class_util_textencoder