@mmoskal 2021-09-21
Just to followup - @pelikhan found that the text encoder was in node since 11.0.0, but I found mkc wasn't exposing it. This is fixed in npm makecode package v0.13.1 https://github.com/microsoft/pxt-mkc/commit/fbfa3a707af5a0c59b919428fe0472b2739a6f32